+++ lib/airtable.js
@@ -13,6 +13,7 @@ export async function airtableGet(path, queryParameters = {}) {
     headers: {
       Authorization: `Bearer ${process.env.AIRTABLE_API_KEY}`,
     },
+    next: { revalidate: 600 },
   });
 
   if (response.ok) {
