+++ lib/api.js
@@ -1,5 +1,5 @@
-import { airtableGet } from './airtable.js';
-import { vtaGet } from './vta.js';
+import { airtableGet } from '#/lib/airtable.js';
+import { vtaGet } from '#/lib/vta.js';
 
 export function fetchAllocations() {
   return airtableGet('Allocations', {
