+++ lib/util.js
@@ -2,12 +2,12 @@
 import querystring from 'querystring';
 import algoliasearch from 'algoliasearch';
 
-import { maxBy, pick, sortBy, uniq } from 'lodash';
+import { maxBy, sortBy, uniq } from 'lodash';
 import bbox from '@turf/bbox';
 import moment from 'moment';
 
-import { categoryCards } from '../lib/category-cards.js';
-import { pageview } from './gtag.js';
+import { categoryCards } from '#/lib/category-cards.js';
+import { pageview } from '#/lib/gtag.js';
 
 const client = algoliasearch(
   process.env.NEXT_PUBLIC_ALGOLIA_APP_ID,
@@ -75,29 +75,29 @@ export function updateUrlWithFilters(filters, currentProjectIds) {
   pageview(newUrl);
 }
 
-export function getFiltersFromQuery(query) {
-  const initialFilters = pick(query, [
-    'transactionType',
-    'grantee',
-    'project',
-    'category',
-  ]);
+export function getFiltersFromQuery(searchParams) {
+  const filters = {};
 
   if (
-    initialFilters.grantee !== '' &&
-    typeof initialFilters.grantee === 'string'
+    searchParams.has('transactionType') &&
+    searchParams.get('transactionType') !== ''
   ) {
-    initialFilters.grantee = [initialFilters.grantee];
+    filters.transactionType = searchParams.get('transactionType');
   }
 
-  if (
-    initialFilters.category !== '' &&
-    typeof initialFilters.category === 'string'
-  ) {
-    initialFilters.category = [initialFilters.category];
+  if (searchParams.has('grantee') && searchParams.get('grantee') !== '') {
+    filters.grantee = searchParams.getAll('grantee');
   }
 
-  return initialFilters;
+  if (searchParams.has('project') && searchParams.get('project') !== '') {
+    filters.project = searchParams.get('project');
+  }
+
+  if (searchParams.has('category') && searchParams.get('category') !== '') {
+    filters.category = searchParams.getAll('category');
+  }
+
+  return filters;
 }
 
 /* eslint-disable-next-line max-params */
