+++ lib/vta.js
@@ -11,6 +11,7 @@ export async function vtaGet(path, queryParameters = {}) {
     headers: {
       'api-key': process.env.VTA_API_KEY,
     },
+    next: { revalidate: 600 },
   });
 
   if (response.ok) {
