+++ package.json
@@ -43,11 +43,13 @@
     "remark-breaks": "^3.0.2"
   },
   "devDependencies": {
+    "@types/node": "18.11.18",
     "eslint": "8.34.0",
     "eslint-config-next": "^13.1.6",
     "husky": "^8.0.3",
     "prettier": "^2.8.4",
-    "pretty-quick": "^3.1.3"
+    "pretty-quick": "^3.1.3",
+    "typescript": "4.9.4"
   },
   "prettier": {
     "singleQuote": true
