--- pages/_app.jsx
@@ -1,65 +1,0 @@
-import 'bootstrap/dist/css/bootstrap.css';
-import '@fortawesome/fontawesome-svg-core/styles.css';
-import 'mapbox-gl/dist/mapbox-gl.css';
-import 'react-slidedown/lib/slidedown.css';
-import '../css/print.css';
-import '../css/index.css';
-import { DefaultSeo } from 'next-seo';
-import HttpsRedirect from 'react-https-redirect';
-import { useEffect } from 'react';
-import Script from 'next/script';
-import { useRouter } from 'next/router';
-import * as gtag from '../lib/gtag';
-
-export default function MyApp({ Component, pageProps }) {
-  const router = useRouter();
-  useEffect(() => {
-    const handleRouteChange = (url) => {
-      gtag.pageview(url);
-    };
-    router.events.on('routeChangeComplete', handleRouteChange);
-    router.events.on('hashChangeComplete', handleRouteChange);
-    return () => {
-      router.events.off('routeChangeComplete', handleRouteChange);
-      router.events.off('hashChangeComplete', handleRouteChange);
-    };
-  }, [router.events]);
-
-  return (
-    <HttpsRedirect>
-      <DefaultSeo
-        openGraph={{
-          type: 'website',
-          locale: 'en_US',
-          url: 'https://2016measureb.vta.org',
-          site_name: '2016 Measure B',
-        }}
-        twitter={{
-          handle: '@VTA',
-          site: '@VTA',
-          cardType: 'summary_large_image',
-        }}
-      />
-      {/* Global Site Tag (gtag.js) - Google Analytics */}
-      <Script
-        strategy="afterInteractive"
-        src={`https://www.googletagmanager.com/gtag/js?id=${gtag.GA_TRACKING_ID}`}
-      />
-      <Script
-        id="gtag-init"
-        strategy="afterInteractive"
-        dangerouslySetInnerHTML={{
-          __html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', '${gtag.GA_TRACKING_ID}', {
-              page_path: window.location.pathname,
-            });
-          `,
-        }}
-      />
-      <Component {...pageProps} />
-    </HttpsRedirect>
-  );
-}
