+++ ui/PrintButton.tsx
@@ -1,10 +1,9 @@
 /* global window */
-import React from 'react';
 import Button from 'react-bootstrap/Button';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPrint } from '@fortawesome/free-solid-svg-icons';
 
-import { event } from '../lib/gtag.js';
+import { event } from '#/lib/gtag.js';
 
 const print = () => {
   event({
@@ -14,12 +13,10 @@ const print = () => {
   window.print();
 };
 
-const PrintButton = ({ className }) => {
+export const PrintButton = ({ className }) => {
   return (
     <Button onClick={print} className={className}>
       <FontAwesomeIcon icon={faPrint} className="mr-2" /> Print
     </Button>
   );
 };
-
-export default PrintButton;
